+++ rundeckapp/grails-app/controllers/rundeck/controllers/ExecutionController.groovy
@@ -25,7 +25,6 @@ import com.dtolabs.rundeck.app.support.ExecutionQuery
 import com.dtolabs.rundeck.app.support.ExecutionQueryException
 import com.dtolabs.rundeck.app.support.ExecutionViewParams
 import com.dtolabs.rundeck.core.authorization.AuthContext
-import com.dtolabs.rundeck.server.projects.AuthContextEvaluatorCacheManager
 import org.rundeck.core.auth.AuthConstants
 import com.dtolabs.rundeck.core.common.PluginDisabledException
 import com.dtolabs.rundeck.core.execution.workflow.state.StateUtils
@@ -70,7 +69,6 @@ class ExecutionController extends ControllerBase{
     FileUploadService fileUploadService
     PluginService pluginService
     ConfigurationService configurationService
-    AuthContextEvaluatorCacheManager authContextEvaluatorCacheManager
 
     static allowedMethods = [
             delete:['POST','DELETE'],
